+++ huey/tests/__init__.py
@@ -8,6 +8,7 @@ from huey.tests.test_immediate import *
 from huey.tests.test_mini import *
 from huey.tests.test_postgres import *
 from huey.tests.test_priority import *
+from huey.tests.test_revoke_atomic import *
 from huey.tests.test_registry import *
 from huey.tests.test_serializer import *
 from huey.tests.test_signals import *